+++ chaco/plots/contour/contour_line_plot.py
