+++ chaco/plots/contour/contour_poly_plot.py
